Converting hostname to lowercase
While .-hostname returns the name in lowercase when testing against
phantom on OSX Yosemite, in El Capitan it does not. This might just
be the older 1.9.8 phantom version I need to use in El Capitan, or it
might be an issue with the version of webkit, which could extend to other
browsers.