No need to re-distribute bootstrap.min.css, we can get it from a CDN
It would make sense to redistribute it if I expected Relevance to work
offline, but given that it's all about tab organization, I think we can
safely assume that either the user is online, or was the last time they
saw the console. This makes the distribution a bit lighter.